Refactor: move UserDefaults extension to UserDefaults+statistic file, add a comma between guessed numbers, divide by parts the project's structure
